requirements.txt
flask
pytesseract
Pillow
opencv-python
numpy